requirements.txt
torch>=1.9.0
esm>=0.4.0
biopython>=1.79
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
requests>=2.25.0
scikit-learn>=1.0.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
omegaconf>=2.3.0